Delay for resize event added, cleanup, data exposure
Moved the resize event handler to the bottom as well as built in a new
internal delay that prevents the event from firing on each dizzang
pixel. Decided to remove the if statement from the resize event because
the wonder of mathemetics. Since a pixel ratio can end up being 1, then
we will just return the value it was * and / with. Added a couple
methods that just expose the canvas style dimensions. Very useful for
accessing when dealing with hi-res canvas.